requirements.txt
streamlit==1.28.1
azure-storage-blob==12.19.0
python-dotenv==1.0.0
PyPDF2==3.0.1
azure-ai-documentintelligence==1.0.0
pandas==2.1.3
openai==1.3.7
requests==2.31.0

# LangChain 관련 (선택사항)
langchain-openai==0.0.5
langchain-community==0.0.10
azure-search-documents==11.4.0 